@@ -1,6 +1,6 @@
-use oxc::{allocator::Allocator, ast::ast::JSXAttributeName, span::Atom};
+use oxc::{ast::ast::JSXAttributeName, span::Atom};
 
-use crate::{analyzer::Analyzer, ty::property_key::PropertyKeyType};
+use crate::{allocator::Allocator, analyzer::Analyzer, ty::property_key::PropertyKeyType};
 
 impl<'a> Analyzer<'a> {
   pub fn exec_jsx_attribute_name(&mut self, node: &'a JSXAttributeName<'a>) -> PropertyKeyType<'a> {
@@ -8,12 +8,12 @@ impl<'a> Analyzer<'a> {
   }
 }
 
-fn get_text<'a>(allocator: &'a Allocator, node: &'a JSXAttributeName<'a>) -> &'a Atom<'a> {
+fn get_text<'a>(allocator: Allocator<'a>, node: &'a JSXAttributeName<'a>) -> &'a Atom<'a> {
   match node {
     JSXAttributeName::Identifier(node) => &node.name,
     JSXAttributeName::NamespacedName(node) => {
-      let s = allocator.alloc(format!("{}:{}", node.namespace.name, node.property.name));
-      allocator.alloc(Atom::from(s.as_str()))
+      let s = allocator.alloc_str(&format!("{}:{}", node.namespace.name, node.name.name));
+      allocator.alloc(Atom::from(&*s))
     }
   }
 }
