@@ -4,7 +4,7 @@ use oxc::ast::{
 
 use crate::{
   analyzer::Analyzer,
-  ty::{property_key::PropertyKeyType, Ty},
+  ty::{Ty, property_key::PropertyKeyType},
 };
 
 impl<'a> Analyzer<'a> {
