@@ -10,7 +10,7 @@ impl<'a> Analyzer<'a> {
 
   pub fn init_binding_identifier(&mut self, node: &'a BindingIdentifier<'a>, init: Option<Ty<'a>>) {
     let symbol = node.symbol_id.get().unwrap();
-    let flags = self.semantic.symbols().get_flags(symbol);
+    let flags = self.semantic.scoping().symbol_flags(symbol);
     let init = if let Some(init) = init {
       self.accumulate_type(node, init);
       init
