@@ -1,6 +1,6 @@
 use oxc::ast::ast::TSInstantiationExpression;
 
-use crate::{ty::Ty, Analyzer};
+use crate::{Analyzer, ty::Ty};
 
 impl<'a> Analyzer<'a> {
   pub fn exec_ts_instantiation_expression(
@@ -9,7 +9,7 @@ impl<'a> Analyzer<'a> {
     _sat: Option<Ty<'a>>,
   ) -> Ty<'a> {
     let base = self.exec_expression(&node.expression, None);
-    let type_args = self.resolve_type_parameter_instantiation(&node.type_parameters);
+    let type_args = self.resolve_type_parameter_instantiation(&node.type_arguments);
     self.instantiate_generic_value(base, &type_args)
   }
 }
