@@ -1,6 +1,6 @@
 use oxc::ast::ast::TSIntersectionType;
 
-use crate::{ty::Ty, Analyzer};
+use crate::{Analyzer, ty::Ty};
 
 impl<'a> Analyzer<'a> {
   pub fn resolve_intersection_type(&mut self, node: &'a TSIntersectionType<'a>) -> Ty<'a> {
