@@ -1,15 +1,20 @@
 use oxc::ast::ast::TSTupleType;
 
 use crate::{
-  ty::{tuple::TupleType, Ty},
   Analyzer,
+  ty::{Ty, tuple::TupleType},
 };
 
 impl<'a> Analyzer<'a> {
   pub fn resolve_tuple_type(&mut self, node: &'a TSTupleType<'a>, readonly: bool) -> Ty<'a> {
-    Ty::Tuple(self.allocator.alloc(TupleType {
-      elements: node.element_types.iter().map(|el| self.resolve_tuple_element(el)).collect(),
-      readonly,
-    }))
+    let allocator = self.allocator;
+    Ty::Tuple(
+      self.allocator.alloc(TupleType {
+        elements: self
+          .allocator
+          .alloc_slice(node.element_types.iter().map(|el| self.resolve_tuple_element(el))),
+        readonly,
+      }),
+    )
   }
 }
