@@ -2,8 +2,8 @@ use std::cell::RefCell;
 
 use oxc::{ast::ast::TSType, semantic::SymbolId, span::Atom};
 
-use super::{ctx::CtxTy, intersection::IntersectionType, union::UnionType, Ty};
-use crate::{analyzer::Analyzer, scope::r#type::TypeScopeId};
+use super::{Ty, ctx::CtxTy, intersection::IntersectionType, record::RecordType, union::UnionType};
+use crate::{allocator, analyzer::Analyzer, scope::r#type::TypeScopeId};
 
 #[derive(Debug, Clone)]
 pub struct GenericParam<'a> {
@@ -18,7 +18,7 @@ pub struct GenericParam<'a> {
 #[derive(Debug, Clone)]
 pub struct GenericType<'a> {
   pub name: &'a Atom<'a>,
-  pub params: Vec<GenericParam<'a>>,
+  pub params: &'a [GenericParam<'a>],
   pub body: CtxTy<'a>,
 }
 
@@ -26,15 +26,15 @@ pub struct GenericType<'a> {
 pub struct GenericInstanceType<'a> {
   pub generic: Ty<'a>,
   /// Defaults should already be applied
-  pub args: Vec<Ty<'a>>,
+  pub args: &'a [Ty<'a>],
   pub unwrapped: RefCell<Option<Ty<'a>>>,
 }
 
 impl<'a> Analyzer<'a> {
   pub fn instantiate_generic_params(
     &mut self,
-    params: &Vec<GenericParam<'a>>,
-    args: &Vec<Ty<'a>>,
+    params: &[GenericParam<'a>],
+    args: &[Ty<'a>],
   ) -> TypeScopeId {
     let scope = self.type_scopes.create_scope();
 
@@ -75,7 +75,7 @@ impl<'a> Analyzer<'a> {
     }
     Ty::Instance(self.allocator.alloc(GenericInstanceType {
       generic,
-      args,
+      args: self.allocator.alloc_slice(args),
       unwrapped: RefCell::new(None),
     }))
   }
@@ -89,13 +89,13 @@ impl<'a> Analyzer<'a> {
 
         // instance.generic is a generic type
         Ty::Generic(generic) => {
-          let scope = self.instantiate_generic_params(&generic.params, &instance.args);
+          let scope = self.instantiate_generic_params(generic.params, instance.args);
           self.resolve_ctx_ty(scope, generic.body)
         }
         Ty::Intrinsic(_) => todo!(),
 
         // instance.generic is a generic value (function or constructor or compound of them)
-        _ => self.instantiate_generic_value(instance.generic, &instance.args),
+        _ => self.instantiate_generic_value(instance.generic, instance.args),
       }
     })
   }
@@ -119,31 +119,34 @@ impl<'a> Analyzer<'a> {
   /// declare const f: (() => string) & (<T>() => T);
   /// const g = f<number>; // Will be `() => number`
   /// ```
-  fn try_instantiate_generic_value(&mut self, ty: Ty<'a>, args: &Vec<Ty<'a>>) -> Option<Ty<'a>> {
+  fn try_instantiate_generic_value(&mut self, ty: Ty<'a>, args: &[Ty<'a>]) -> Option<Ty<'a>> {
     match ty {
       Ty::Function(f) => self.instantiate_callable_type_parameters(f, args).map(Ty::Function),
       Ty::Constructor(c) => self.instantiate_callable_type_parameters(c, args).map(Ty::Constructor),
 
       Ty::Union(union) => {
-        let complex =
-          union.complex.iter().map(|ty| self.instantiate_generic_value(*ty, args)).collect();
-        Some(Ty::Union(self.allocator.alloc(Box::new(UnionType { complex, ..union.clone() }))))
+        let mut complex = allocator::HashMap::new_in(self.allocator);
+        for ty in union.complex.keys() {
+          complex.insert(self.instantiate_generic_value(*ty, args), ());
+        }
+        Some(Ty::Union(self.allocator.alloc(UnionType { complex, ..union.clone() })))
       }
       Ty::Intersection(intersection) => {
         let kind = intersection.kind;
-        let object_like: Vec<_> = intersection
-          .object_like
-          .iter()
-          .filter_map(|ty| self.try_instantiate_generic_value(*ty, args))
-          .collect();
+        let object_like = self.allocator.alloc_slice(
+          intersection
+            .object_like
+            .iter()
+            .filter_map(|ty| self.try_instantiate_generic_value(*ty, args)),
+        );
         if object_like.is_empty() {
           intersection.kind_to_ty()
         } else {
-          Some(Ty::Intersection(self.allocator.alloc(Box::new(IntersectionType {
+          Some(Ty::Intersection(self.allocator.alloc(IntersectionType {
             kind,
             object_like,
-            unresolved: intersection.unresolved.clone(),
-          }))))
+            unresolved: intersection.unresolved,
+          })))
         }
       }
 
@@ -152,16 +155,16 @@ impl<'a> Analyzer<'a> {
         Some(self.instantiate_generic_value(unwrapped, args))
       }
 
-      Ty::Unresolved(_) => Some(self.create_generic_instance(ty, args.clone())),
+      Ty::Unresolved(_) => Some(self.create_generic_instance(ty, args.to_vec())),
 
       ty => Some(ty),
     }
   }
 
-  pub fn instantiate_generic_value(&mut self, ty: Ty<'a>, args: &Vec<Ty<'a>>) -> Ty<'a> {
+  pub fn instantiate_generic_value(&mut self, ty: Ty<'a>, args: &[Ty<'a>]) -> Ty<'a> {
     self
       .try_instantiate_generic_value(ty, args)
-      .unwrap_or_else(|| Ty::Record(self.allocator.alloc(Default::default())))
+      .unwrap_or_else(|| Ty::Record(self.allocator.alloc(RecordType::empty_in(self.allocator))))
   }
 
   pub fn serialize_instance_type(&mut self, instance: &GenericInstanceType<'a>) -> TSType<'a> {
